Adds tests against PHPUnit 10.5
