+++ src/Illuminate/Testing/Constraints/SeeInOrder.php
@@ -3,86 +3,16 @@
 namespace Illuminate\Testing\Constraints;
 
 use PHPUnit\Framework\Constraint\Constraint;
-use ReflectionClass;
+use PHPUnit\Runner\Version;
 
-readonly class SeeInOrder extends Constraint
-{
-    /**
-     * The string under validation.
-     *
-     * @var string
-     */
-    protected string $content;
-
-    /**
-     * The last value that failed to pass validation.
-     *
-     * @var string
-     */
-    protected string $failedValue;
-
-    /**
-     * Create a new constraint instance.
-     *
-     * @param  string  $content
-     * @return void
-     */
-    public function __construct($content)
+if (str_starts_with(Version::series(), '10')) {
+    class SeeInOrder extends Constraint
     {
-        $this->content = $content;
+        use Concerns\SeeInOrder;
     }
-
-    /**
-     * Determine if the rule passes validation.
-     *
-     * @param  array  $values
-     * @return bool
-     */
-    public function matches($values): bool
+} else {
+    readonly class SeeInOrder extends Constraint
     {
-        $position = 0;
-
-        foreach ($values as $value) {
-            if (empty($value)) {
-                continue;
-            }
-
-            $valuePosition = mb_strpos($this->content, $value, $position);
-
-            if ($valuePosition === false || $valuePosition < $position) {
-                $this->failedValue = $value;
-
-                return false;
-            }
-
-            $position = $valuePosition + mb_strlen($value);
-        }
-
-        return true;
-    }
-
-    /**
-     * Get the description of the failure.
-     *
-     * @param  array  $values
-     * @return string
-     */
-    public function failureDescription($values): string
-    {
-        return sprintf(
-            'Failed asserting that \'%s\' contains "%s" in specified order.',
-            $this->content,
-            $this->failedValue
-        );
-    }
-
-    /**
-     * Get a string representation of the object.
-     *
-     * @return string
-     */
-    public function toString(): string
-    {
-        return (new ReflectionClass($this))->name;
+        use Concerns\SeeInOrder;
     }
 }
